refactor: 의미가 모호한 Dto를 Domain으로 전환
-Profit과 MatchCount는 요구사항에 관한 것으로 Domain으로 전환한다.
