+++ src/main/java/lotto/domain/MatchCount.java
@@ -1,0 +1,4 @@
+package lotto.domain;
+
+public record MatchCount(int matchCount, boolean bonus) {
+}
