+++ src/main/java/lotto/service/LottoService.java
@@ -8,16 +8,16 @@ import java.util.Map;
 
 import lotto.domain.Amount;
 import lotto.domain.MatchStatistics;
-import lotto.dto.MatchCountDto;
-import lotto.dto.Profit;
+import lotto.domain.MatchCount;
+import lotto.domain.Profit;
 
 public class LottoService {
     private final static double PROFIT_BENCHMARK = 1.0;
 
-    public Map<MatchStatistics, Integer> convertToMap(List<MatchCountDto> matchCount) {
+    public Map<MatchStatistics, Integer> convertToMap(List<MatchCount> matchCount) {
         Map<MatchStatistics, Integer> map = new HashMap<>();
 
-        for (MatchCountDto dto : matchCount) {
+        for (MatchCount dto : matchCount) {
             MatchStatistics statistic = getMatchStatistics(dto.matchCount(), dto.bonus());
             map.put(statistic, map.getOrDefault(statistic, 0) + 1);
         }
@@ -34,7 +34,7 @@ public class LottoService {
             sum += matchMoney * count;
         }
 
-        double result = (double)sum / amount.getMoney();
+        double result = (double)sum / amount.money();
         return new Profit(result, isProfit(result));
     }
